PEZY-497: Build 4 stat cards
